requirements.txt
# SetupSpawn Shorts Scraper Dependencies

# Video discovery and download
yt-dlp>=2023.12.30

# Transcription services
youtube-transcript-api>=0.6.2
assemblyai>=0.21.0

# Video/image processing
ffmpeg-python>=0.2.0
opencv-python>=4.8.0

# AI/LLM services
openai>=1.6.0

# CLI framework
typer>=0.9.0

# Configuration
pyyaml>=6.0.1

# Logging
structlog>=23.2.0

# Data processing
pandas>=2.0.0

# Testing
pytest>=7.4.0

# Development tools
black>=23.0.0
flake8>=6.0.0
